+++ Day17/Day17/Tools/Math/MathHelper.cs
@@ -2,7 +2,7 @@ namespace Day17.Tools.Math
 {
     public static class MathHelper
     {
-        public static long ManhattanDistance( IPlaceable left, IPlaceable right )
+        public static int ManhattanDistance( IPlaceable left, IPlaceable right )
         {
             return System.Math.Abs( left.CoordinateRow - right.CoordinateRow ) +
                    System.Math.Abs( left.CoordinateColumn - right.CoordinateColumn );
